adding Target and Hash obj to pub def + del unnecessary includes
