+++ src/libaktualizr/primary/sotauptaneclient.h
@@ -28,7 +28,6 @@
 #include "uptane/fetcher.h"
 #include "uptane/imagerepository.h"
 #include "uptane/iterator.h"
-#include "uptane/tuf.h"
 
 class SotaUptaneClient {
  public:
